+++ CHANGES.md
@@ -8,6 +8,9 @@ Version 1.3.2
 
 To be released.
 
+ -  Fixed the default document loader to handle the `Link` header with
+    incorrect syntax.  [[#196]]
+
 
 Version 1.3.1
 -------------
@@ -133,6 +136,15 @@ Released on November 30, 2024.
 [#186]: https://github.com/dahlia/fedify/pull/186
 [#192]: https://github.com/dahlia/fedify/issues/192
 [#193]: https://github.com/dahlia/fedify/issues/193
+
+
+Version 1.2.10
+--------------
+
+Released on December 18, 2024.
+
+ -  Fixed the default document loader to handle the `Link` header with
+    incorrect syntax.  [[#196]]
 
 
 Version 1.2.9
@@ -348,6 +360,15 @@ Released on October 31, 2024.
 
 [implicit contexts]: https://logtape.org/manual/contexts#implicit-contexts
 [#118]: https://github.com/dahlia/fedify/issues/118
+
+
+Version 1.1.10
+--------------
+
+Released on December 18, 2024.
+
+ -  Fixed the default document loader to handle the `Link` header with
+    incorrect syntax.  [[#196]]
 
 
 Version 1.1.9
@@ -604,6 +625,17 @@ Released on October 20, 2024.
 [ActivityPub and HTTP Signatures]: https://swicg.github.io/activitypub-http-signature/
 [#146]: https://github.com/dahlia/fedify/issues/146
 [#150]: https://github.com/dahlia/fedify/issues/150
+
+
+Version 1.0.13
+--------------
+
+Released on December 18, 2024.
+
+ -  Fixed the default document loader to handle the `Link` header with
+    incorrect syntax.  [[#196]]
+
+[#196]: https://github.com/dahlia/fedify/issues/196
 
 
 Version 1.0.12